Move styleguide route to /app for reuse in subtheme
This moves the code for the styleguide route to the /app directory so we
can easily mount the styleguide inside a subtheme's dummy app.

Eventually we should instead only do this in development or allow it to
be configured by the user but this works for now and makes it much
easier to work on the subtheme.

This also moves a few `devDependencies` to `dependencies` so that
they're available in the subtheme.